Add ngnix container Dockerfile
Add dockerfile to create nginx instance and replace the default config with our nginx.conf file.

Note the config directories:
In a traditional web server, nginx config files are stored in /sites-enabled/ in the apache style.
In docker, nginx configs are stored in conf.d, which is the more correct way in nginx.